Rename repositories to kebab-case with type extension
