+++ src/controllers/company-ch.controller.js
@@ -3,7 +3,7 @@ const express = require('express')
 const winston = require('winston')
 const Q = require('q')
 const { getFormattedAddress } = require('../lib/address')
-const companyRepository = require('../repositorys/companyrepository')
+const companyRepository = require('../repos/company.repo')
 const companyFormattingService = require('../services/company-formatting.service')
 const { chDetailsLabels } = require('../labels/companylabels')
