+++ src/controllers/company-ltd.controller.js
@@ -5,8 +5,8 @@ const Q = require('q')
 const companyService = require('../services/company.service')
 const companyFormService = require('../services/company-form.service')
 const companyFormattingService = require('../services/company-formatting.service')
-const companyRepository = require('../repositorys/companyrepository')
-const metadataRepository = require('../repositorys/metadatarepository')
+const companyRepository = require('../repos/company.repo')
+const metadataRepository = require('../repos/metadata.repo')
 const { companyDetailsLabels, chDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels/companylabels')
 const { genCSRF, containsFormData, isBlank } = require('../lib/controllerutils')
 const router = express.Router()
