+++ src/controllers/company-ukother.controller.js
@@ -2,12 +2,12 @@
 const express = require('express')
 const winston = require('winston')
 const Q = require('q')
-const companyRepository = require('../repositorys/companyrepository')
+const companyRepository = require('../repos/company.repo')
 const companyFormService = require('../services/company-form.service')
 const companyService = require('../services/company.service')
 const companyFormattingService = require('../services/company-formatting.service')
 const { companyDetailsLabels, accountManagementDisplayLabels, hqLabels } = require('../labels/companylabels')
-const metadataRepository = require('../repositorys/metadatarepository')
+const metadataRepository = require('../repos/metadata.repo')
 const { genCSRF, containsFormData, isBlank } = require('../lib/controllerutils')
 const router = express.Router()
 const companyWithoutCHKeys = ['business_type', 'registered_address', 'alias', 'trading_address', 'uk_region', 'headquarter_type', 'sector', 'website', 'description', 'employee_range', 'turnover_range']
