+++ src/controllers/contact-edit.controller.js
@@ -4,8 +4,8 @@ const winston = require('winston')
 const { genCSRF, containsFormData } = require('../lib/controllerutils')
 const contactFormService = require('../services/contact-form.service')
 const { contactLabels } = require('../labels/contactlabels')
-const metadataRepository = require('../repositorys/metadatarepository')
-const companyRepository = require('../repositorys/companyrepository')
+const metadataRepository = require('../repos/metadata.repo')
+const companyRepository = require('../repos/company.repo')
 const { getCommon } = require('../controllers/contact.controller')
 
 const router = express.Router()
