+++ src/controllers/contact.controller.js
@@ -3,7 +3,7 @@ const winston = require('winston')
 const Q = require('q')
 
 const { genCSRF } = require('../lib/controllerutils')
-const contactRepository = require('../repositorys/contactrepository')
+const contactRepository = require('../repos/contact.repo')
 const contactFormattingService = require('../services/contact-formatting.service')
 const companyService = require('../services/company.service')
 const { contactDetailsLabels } = require('../labels/contactlabels')
