+++ src/controllers/interaction-edit.controller.js
@@ -2,10 +2,10 @@
 const express = require('express')
 const Q = require('q')
 const interactionLabels = require('../labels/interactionlabels')
-const contactRepository = require('../repositorys/contactrepository')
-const companyRepository = require('../repositorys/companyrepository')
-const metadataRepository = require('../repositorys/metadatarepository')
-const advisorRepository = require('../repositorys/advisorrepository')
+const contactRepository = require('../repos/contact.repo')
+const companyRepository = require('../repos/company.repo')
+const metadataRepository = require('../repos/metadata.repo')
+const advisorRepository = require('../repos/advisor.repo')
 const interactionDataService = require('../services/interaction-data.service')
 const interactionFormService = require('../services/interaction-form.service')
 const { genCSRF, containsFormData } = require('../lib/controllerutils')
