+++ src/controllers/service-delivery.controller.js
@@ -5,8 +5,8 @@ const Q = require('q')
 const serviceDeliverylabels = require('../labels/servicedelivery')
 const { genCSRF, transformV2Errors } = require('../lib/controllerutils')
 const { nullEmptyFields, deleteNulls } = require('../lib/propertyhelpers')
-const metadataRepository = require('../repositorys/metadatarepository')
-const serviceDeliveryRepository = require('../repositorys/servicedeliveryrepository')
+const metadataRepository = require('../repos/metadata.repo')
+const serviceDeliveryRepository = require('../repos/service-delivery.repo')
 const serviceDeliveryService = require('../services/service-delivery.service')
 const { getDisplayServiceDelivery } = require('../services/service-delivery-formatting.service')
 const { getViewCompanyLink } = require('../services/company.service')
