+++ src/server.js
@@ -16,7 +16,7 @@ const filters = require('@uktrade/trade_elements/dist/nunjucks/filters')
 const forceHttps = require('./middleware/forcehttps')
 const headers = require('./middleware/headers')
 const locals = require('./middleware/locals')
-const metadata = require('./repositorys/metadatarepository')
+const metadata = require('./repos/metadata.repo')
 const user = require('./middleware/user')
 const auth = require('./middleware/auth')
 const csrf = require('./middleware/csrf')
