+++ src/services/company.service.js
@@ -1,10 +1,10 @@
 /* eslint camelcase: 0 */
 const Q = require('q')
 const winston = require('winston')
-const advisorRepository = require('../repositorys/advisorrepository')
-const companyRepository = require('../repositorys/companyrepository')
-const metadataRepository = require('../repositorys/metadatarepository')
-const serviceDeliveryRepository = require('../repositorys/servicedeliveryrepository')
+const advisorRepository = require('../repos/advisor.repo')
+const companyRepository = require('../repos/company.repo')
+const metadataRepository = require('../repos/metadata.repo')
+const serviceDeliveryRepository = require('../repos/service-delivery.repo')
 const interactionDataService = require('./interaction-data.service')
 const { genCSRF } = require('../lib/controllerutils')
 const { getFormattedAddress } = require('../lib/address')
