+++ test/controllers/search.controller.test.js
@@ -8,7 +8,7 @@ describe('Search controller', function () {
   describe('view company result', function () {
     it('should route a ch result', function (done) {
       const searchController = proxyquire('../../src/controllers/search.controller', {
-        '../repositorys/companyrepository': {
+        '../repos/company.repo': {
           getDitCompany: sinon.stub().resolves(null)
         }
       })
@@ -30,7 +30,7 @@ describe('Search controller', function () {
     })
     it('should route a uk private ltd company', function (done) {
       const searchController = proxyquire('../../src/controllers/search.controller', {
-        '../repositorys/companyrepository': {
+        '../repos/company.repo': {
           getDitCompany: sinon.stub().resolves({
             id: '9999',
             uk_based: true,
@@ -60,7 +60,7 @@ describe('Search controller', function () {
     })
     it('should route a uk public ltd company', function (done) {
       const searchController = proxyquire('../../src/controllers/search.controller', {
-        '../repositorys/companyrepository': {
+        '../repos/company.repo': {
           getDitCompany: sinon.stub().resolves({
             id: '9999',
             uk_based: true,
@@ -90,7 +90,7 @@ describe('Search controller', function () {
     })
     it('should route a uk public other company', function (done) {
       const searchController = proxyquire('../../src/controllers/search.controller', {
-        '../repositorys/companyrepository': {
+        '../repos/company.repo': {
           getDitCompany: sinon.stub().resolves({
             id: '9999',
             uk_based: true,
@@ -120,7 +120,7 @@ describe('Search controller', function () {
     })
     it('should route a foreign company', function (done) {
       const searchController = proxyquire('../../src/controllers/search.controller', {
-        '../repositorys/companyrepository': {
+        '../repos/company.repo': {
           getDitCompany: sinon.stub().resolves({
             id: '9999',
             uk_based: false,
